/home/fran/Descargas/ENTER/bin/python /home/fran/Documentos/CUARTO_CARRERA/ALT/SAR_WEB_CRAWLER/ALT_BusquedaAproximada/testing.py

+++ ALT_BusquedaAproximada/distancias.py
@@ -4,7 +4,7 @@ def levenshtein_matriz(x, y, threshold=None):
     # esta versión no utiliza threshold, se pone porque se puede
     # invocar con él, en cuyo caso se ignora
     lenX, lenY = len(x), len(y)
-    D = np.zeros((lenX + 1, lenY + 1), dtype=np.int)
+    D = np.zeros((lenX + 1, lenY + 1))
     for i in range(1, lenX + 1):
         D[i][0] = D[i - 1][0] + 1
     for j in range(1, lenY + 1):
